requirements.txt
markdown
django-filter
djangorestframework
django==4.2
dj-rest-auth
dj-rest-auth[with_social]
djangorestframework-simplejwt
django-allauth[socialaccount]
Pillow
django-cors-headers